Consolidate forward declarations into per-subdirectory Fwd.hpp files
Replace scattered inline forward declarations with centralized Fwd.hpp
includes across collision, constraint, dynamics, gui, simulation, and
utils subdirectories to improve maintainability and reduce duplication.

- Create dart/gui/Fwd.hpp with forward declarations for gui classes
- Create dart/utils/Fwd.hpp with forward declarations for utils classes
- Add VisualAspect to dart/dynamics/Fwd.hpp
- Add Recording to dart/simulation/Fwd.hpp
- Update 21 header files to use Fwd.hpp includes instead of inline
  forward declarations
